move peer.ts
